Spli battle and inventory UI logic into separate scripts, and move files into separate folders.
