+++ Assets/Scripts/Battle/BattleUIScript.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 47db8e374adc6fa42a98147d65792a41
+guid: c0e18f15fad9ea04eb0951e7967a43bb
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
