+++ Assets/Scripts/Battle/BeamScript.cs
