+++ Assets/Scripts/Battle/FillingBarScript.cs
