+++ Assets/Scripts/Inventory.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 13cdc1269a5bb8c4284f99c3c846bea3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
