+++ Assets/Scripts/Inventory/InventoryUIScript.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 47db8e374adc6fa42a98147d65792a41
+guid: cf5645c80982a8342901a241f380f4d0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
